schoology-scraper/requirements.txt
# Scraper-specific dependencies
selenium==4.16.0
beautifulsoup4==4.12.3
boto3==1.34.34
python-dotenv==1.0.1
deepdiff
google.generativeai

# Web scraping dependencies
requests==2.31.0
lxml==5.3.0
certifi==2023.11.17
charset-normalizer==3.3.2

# Selenium WebDriver dependencies
h11==0.14.0
outcome==1.3.0.post0
PySocks==1.7.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
websockets==12.0
wsproto==1.2.0

# Utilities
attrs==23.2.0
soupsieve==2.6
idna==3.6
setuptools